+++ src/renderer/src/components/client/ClientGrids/RegisterGrid/RegisterGridToolbar/MenuButton/MenuButton.tsx
@@ -2,6 +2,7 @@ import { useScanRegistersZustand } from '@renderer/components/client/ClientGrids
 import { meme } from '@renderer/components/shared/inputs/meme'
 import { useRef, useState } from 'react'
 import LoadDummyDataButton from './LoadDummyDataButton/LoadDummyDataButton'
+import MenuConnectionOptions from './MenuConnectionOptions/MenuConnectionOptions'
 import MenuRegisterOptions from './MenuRegisterOptions/MenuRegisterOptions'
 import ScanRegistersButton, { SetAnchorProps } from './ScanRegistersButton/ScanRegistersButton'
 import ScanUnitIds from './ScanUnitIds/ScanUnitIds'
@@ -14,6 +15,7 @@ const MenuContent = meme(({ setAnchor }: SetAnchorProps) => {
   return (
     <FormGroup>
       <MenuRegisterOptions />
+      <MenuConnectionOptions />
       <ScanUnitIds />
       <ScanRegistersButton setAnchor={setAnchor} />
       <LoadDummyDataButton setAnchor={setAnchor} />